+++ src/allocation/domain/pano_fdi_segmentation/schemas.py
@@ -22,7 +22,7 @@ class CvatSegmentation(BaseModel):
     confidence: float
     label: str
     type: str
-    points: List[float]
+    points: List[int]
     #mask: List[int]
 
 class CvatSegmentations(BaseModel):
